changed to dynamic routing and layout
