+++ views/content/bloop/blop.ejs
@@ -1,0 +1,4 @@
+<section id="home">
+    <hr>
+    <h1 id="name">blop</h1>
+</section>
