+++ views/content/index.ejs
@@ -1,0 +1,51 @@
+<!--Home-->
+
+<section id="home">
+    <hr>
+    <h1 id="name">Til Georg Pockrandt</h1>
+    <h2>Schüler am <a href="bszet.de">BSZ für Elektrotechnik</a> <br><b>Diese Seite ist noch in Entwicklung und noch nicht fertig</b></h2>
+    <a href="#about">
+        <i class="fas fa-arrow-alt-circle-down fa-5x" id="scrollToBottomBtn"></i>
+    </a>
+    <i onclick="topFunction()" class="fas fa-arrow-alt-circle-up fa-5x" id="scrollToTopBtn"></i>
+</section>
+
+<!--About-->
+
+<section id="about">
+    <h3>Über mich</h3>
+    <hr>
+    <img src="static/img/avatar.png" alt="Avatar">
+    <h4>Til Georg Pockrandt</h4>
+    <p>Hallo, Ich bin 17 Jahre alt und das ist meine Webseite. <br> Es ist noch nicht sehr viel hier, aber ich werde nach und nach Projekte hier dokumentieren.</p>
+</section>
+
+<!--Projects-->
+
+<section id="projects">
+    <h3>Projekte</h3>
+    <hr>
+    <div>
+        <ul>
+            <li><a href="https://github.com/TilGP/website"><img src="https://via.placeholder.com/280x170" alt="Placeholder_IMG_280x170"></a></li>
+            <li><a href="#"><img src="https://via.placeholder.com/280x170" alt="Placeholder_IMG_280x170"></a></li>
+            <li><a href="#"><img src="https://via.placeholder.com/280x170" alt="Placeholder_IMG_280x170"></a></li>
+            <li><a href="#"><img src="https://via.placeholder.com/280x170" alt="Placeholder_IMG_280x170"></a></li>
+            <li><a href="#"><img src="https://via.placeholder.com/280x170" alt="Placeholder_IMG_280x170"></a></li>
+            <li><a href="#"><img src="https://via.placeholder.com/280x170" alt="Placeholder_IMG_280x170"></a></li>
+        </ul>
+    </div>
+</section>
+
+<!--Contact-->
+
+<section id="contact">
+    <h3>Kontakt</h3>
+    <hr>
+
+    <ul>
+        <li>E-Mail: <a href="mailto:tp@til-pockrandt.de">tp@til-pockrandt.de</a></li>
+        <li>Github: <a href="https://github.com/TilGP" target="_blank">TilGP</a></li>
+    </ul>
+    <script src="static/scripts/scrollToTop.js"></script>
+</section>
