+++ views/content/test2.ejs
@@ -1,0 +1,8 @@
+<section id="home">
+    <hr>
+    <h1 id="name">Testpage two</h1>
+    <h2>Schüler am <a href="bszet.de">BSZ für Elektrotechnik</a> <br><b>Diese Seite ist noch in Entwicklung und noch nicht fertig</b></h2>
+    <a href="#about">
+        <i class="fas fa-arrow-alt-circle-down fa-5x" id="scrollToBottomBtn"></i>
+    </a>
+</section>
