+++ views/footer.ejs
@@ -3,5 +3,3 @@
         &copy; 2021 - Til Georg Pockrandt
     </p>
 </footer>
-<script src="scripts/navbar.js"></script>
-<script src="scripts/scrollToTop.js"></script>
