+++ views/header.ejs
@@ -2,7 +2,7 @@
 <header>
     <div id="logo">
         <a href="#">
-            <img src="img/logo.png" alt="Logo">
+            <img src="static/img/logo.png" alt="Logo">
         </a>
     </div>
