+++ views/layout.ejs
@@ -1,0 +1,25 @@
+<!doctype html>
+<html lang="de">
+ <head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta name="robots" content="noindex,follow" />
+  <link href="https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@400&family=Roboto:wght@500&display=swap" rel="stylesheet">
+  <link rel="stylesheet" type="text/css" href="static/css/style.css">
+  <script src="https://kit.fontawesome.com/8b368c47c1.js" crossorigin="anonymous"></script>
+  <title>Site is still under construction!</title>
+ </head>
+ <body>
+
+    <!--Header-->
+    <%- include("header") %>
+
+    <!--Content-->
+    <%- include("./content" + url) %>
+
+    <!--Footer-->
+    <%- include("footer") %>
+    
+    <script src="static/scripts/navbar.js"></script>
+ </body>
+</html>
